% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTMLE-package.R
\docType{package}
\name{cleanTMLE-package}
\alias{cleanTMLE}
\alias{cleanTMLE-package}
\title{cleanTMLE: Outcome-Blind Design and Estimand Selection for Targeted Learning}
\description{
Declares an analysis plan, locks the design data without the outcome, checks by outcome-free plasmode simulation which estimands the design can deliver and which targeted maximum likelihood candidate to use for each, records that decision in a hashed dossier, and only then admits the outcome for estimation. Weighting and matching are offered as comparators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amertens/cleanTMLE}
  \item Report bugs at \url{https://github.com/amertens/cleanTMLE/issues}
}

}
\author{
\strong{Maintainer}: Andrew Mertens \email{amertens@berkeley.edu}

Other contributors:
\itemize{
  \item cleanTMLE Contributors [contributor]
}

}
\keyword{internal}
