% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.apply_weight_constraints}
\alias{.apply_weight_constraints}
\title{Apply trimming and truncation to weights}
\usage{
.apply_weight_constraints(
  wt_result,
  trim = NULL,
  trunc = NULL,
  data = NULL,
  trt_vals = NULL
)
}
\description{
Apply trimming and truncation to weights
}
\keyword{internal}
