% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ipwrisk.R
\name{.attach_boot_ci}
\alias{.attach_boot_ci}
\title{Attach bootstrap CIs to risk data.frame}
\usage{
.attach_boot_ci(risk_df, boot_results)
}
\description{
Attach bootstrap CIs to risk data.frame
}
\keyword{internal}
