% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ipwrisk.R
\name{.bootstrap_ipwrisk}
\alias{.bootstrap_ipwrisk}
\title{Bootstrap helper for IPW risk}
\usage{
.bootstrap_ipwrisk(spec, nboot, risk_time, trim, trunc, weight_type)
}
\description{
Bootstrap helper for IPW risk
}
\keyword{internal}
