% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ipwrisk.R
\name{.build_event_indicator}
\alias{.build_event_indicator}
\title{Build event indicator vector}
\usage{
.build_event_indicator(data, spec)
}
\description{
Build event indicator vector
}
\keyword{internal}
