% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.build_table1}
\alias{.build_table1}
\title{Internal Table 1 builder}
\usage{
.build_table1(data, vars = NULL, by = NULL, weights = NULL, smd = TRUE)
}
\description{
Internal Table 1 builder
}
\keyword{internal}
