% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_clean_room_wrapper.R
\name{.build_weight_summary}
\alias{.build_weight_summary}
\title{Build weight summary table from raw vectors}
\usage{
.build_weight_summary(w, A)
}
\description{
Build weight summary table from raw vectors
}
\keyword{internal}
