% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.compute_contrasts}
\alias{.compute_contrasts}
\title{Compute risk contrasts (RD or RR) between two groups}
\usage{
.compute_contrasts(risk_df, measure = "RD")
}
\description{
Compute risk contrasts (RD or RR) between two groups
}
\keyword{internal}
