% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{.surv_tmle_via_package}
\alias{.surv_tmle_via_package}
\title{survtmle package wrapper}
\usage{
.surv_tmle_via_package(
  data,
  treatment,
  time,
  event,
  covariates,
  target_times,
  sl_library,
  ...
)
}
\description{
survtmle package wrapper
}
\keyword{internal}
