% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{load_audit}
\alias{load_audit}
\title{Load an Audit Log from Disk}
\usage{
load_audit(path)
}
\arguments{
\item{path}{Character; path to the RDS file.}
}
\value{
A \code{cleantmle_audit} object.
}
\description{
Load an Audit Log from Disk
}
\keyword{internal}
