% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_glmnet_bounded.R
\name{predict.SL.glmnet.bounded}
\alias{predict.SL.glmnet.bounded}
\title{Predict Method for SL.glmnet.bounded}
\usage{
\method{predict}{SL.glmnet.bounded}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{SL.glmnet.bounded} fit.}

\item{newdata}{Predictor data.frame.}

\item{...}{Unused.}
}
\value{
A numeric vector of predicted responses.
}
\description{
Predict Method for SL.glmnet.bounded
}
