% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{re_estimate}
\alias{re_estimate}
\title{Re-estimate a Fitted Model}
\usage{
re_estimate(x, ...)
}
\arguments{
\item{x}{A fitted result object whose spec may have been updated.}

\item{...}{Additional arguments passed to the estimation function.}
}
\value{
A new fitted result object.
}
\description{
Re-runs the estimation function that produced the fitted object,
using the (potentially updated) specification.
}
