% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_treatment}
\alias{update_treatment}
\alias{update_outcome}
\alias{update_censoring}
\alias{re_estimate}
\alias{compare_fits}
\title{Update and Re-estimate Model Components}
\usage{
update_treatment(x, ...)
update_outcome(x, ...)
update_censoring(x, ...)
re_estimate(x, ...)
compare_fits(..., risk_time = NULL)
}
\arguments{
\item{x}{A \code{cr_spec} or fitted result object.}
\item{risk_time}{For compare_fits: time for comparison.}
\item{...}{Arguments passed to identify_* or estimation functions,
  or named fitted objects for compare_fits.}
}
\description{
Modify specification components and re-fit models.
}
