% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{verify_lock}
\alias{verify_lock}
\title{Verify a lock's hash}
\usage{
verify_lock(lock)
}
\arguments{
\item{lock}{A \code{cr_lock}.}
}
\value{
\code{TRUE}, invisibly; an error if anything locked has changed.
}
\description{
Verify a lock's hash
}
