% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmode_dq.R
\name{.degrade_missingness}
\alias{.degrade_missingness}
\title{Introduce MCAR missingness into covariates}
\usage{
.degrade_missingness(data, covariates, fraction, seed_offset)
}
\description{
Introduce MCAR missingness into covariates
}
\keyword{internal}
