% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmode_dq.R
\name{.degrade_outcome}
\alias{.degrade_outcome}
\title{Apply outcome misclassification}
\usage{
.degrade_outcome(Y, sensitivity, specificity)
}
\description{
Apply outcome misclassification
}
\keyword{internal}
