% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmode_dq.R
\name{.degrade_treatment}
\alias{.degrade_treatment}
\title{Flip treatment assignment (non-differential misclassification)}
\usage{
.degrade_treatment(A, rate)
}
\description{
Flip treatment assignment (non-differential misclassification)
}
\keyword{internal}
