% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{.surv_tmle_fallback}
\alias{.surv_tmle_fallback}
\title{Fallback survival TMLE via binarization}
\usage{
.surv_tmle_fallback(
  data,
  treatment,
  time,
  event,
  covariates,
  target_times,
  sl_library,
  ...
)
}
\description{
Fallback survival TMLE via binarization
}
\keyword{internal}
