% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{get_final_cohort}
\alias{get_final_cohort}
\title{Extract the Final Cohort from an Attrition Table}
\usage{
get_final_cohort(attrition)
}
\arguments{
\item{attrition}{A \code{cleantmle_attrition} object.}
}
\value{
A data.frame.
}
\description{
Returns the data.frame after all inclusion/exclusion criteria have
been applied.
}
