% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_clean_room_wrapper.R
\name{ic_histogram}
\alias{ic_histogram}
\title{Influence Curve Histogram}
\usage{
ic_histogram(tmle_result, bins = 30L)
}
\arguments{
\item{tmle_result}{A \code{tmle_fit} object with an
\code{influence_curve} element.}

\item{bins}{Integer; number of histogram bins. Default: 30.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots a histogram of the efficient influence curve (EIC) values from
a fitted TMLE object.
}
