% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{summarize_plasmode_results}
\alias{summarize_plasmode_results}
\title{Summarize Plasmode Results}
\usage{
summarize_plasmode_results(x, ...)
}
\arguments{
\item{x}{A \code{plasmode_results} object from \code{\link[=run_plasmode_feasibility]{run_plasmode_feasibility()}}.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Convenience function that prints and invisibly returns the performance
metrics from \code{\link[=run_plasmode_feasibility]{run_plasmode_feasibility()}}.
}
