% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_methods}
\alias{update_methods}
\alias{update_treatment}
\alias{update_outcome}
\alias{update_censoring}
\title{Update and Re-estimate Model Components}
\usage{
update_treatment(x, ...)

update_outcome(x, ...)

update_censoring(x, ...)
}
\arguments{
\item{x}{A \code{cr_spec} or fitted result object.}

\item{...}{Arguments passed to the corresponding \verb{identify_*} function.}
}
\description{
These functions modify an existing specification or fitted object's
treatment, outcome, or censoring specification and optionally re-fit.
}
